==> 1_cleaning_data.ipynb <==
### Import Dependencies

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

### Loading the Data

df = pd.read_csv('data/raw/dataset.csv')
pd.set_option('display.max_columns', None)
df.head()

df.shape

df.dtypes

df.isnull().sum()

There are no missing values

### Checking for Inconsistencies

for col in df.select_dtypes(include=['object']).columns:
    print(f'{col}: {df[col].nunique()} unique values')
    if col not in ['customerID', 'TotalCharges']:
        for value in df[col].unique():
            print(f' - {value}')

### Data Cleaning

# Convert TotalCharges column to float
df['TotalCharges'] = pd.to_numeric(df['TotalCharges'], errors='coerce')
df['Churn'] = df['Churn'].map({'No': 0, 'Yes': 1})

df

df.drop(columns=['customerID'], inplace=True)

df

df.to_csv('data/processed/1_data_cleaned.csv', index=False)
