==> 4_feature_engineering.ipynb <==
### Import Dependencies

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.base import BaseEstimator, TransformerMixin

### Loading the Data

df = pd.read_csv('data/processed/3_outliers_handled.csv')

### Feature Engineering

#### Service Adoption Score

services = ['OnlineSecurity', 'OnlineBackup', 'DeviceProtection', 'TechSupport', 'StreamingTV', 'StreamingMovies']
services_binary = df[services].replace({'Yes': 1, 'No': 0, 'No internet service': 0})

df['ServiceAdoptionScore'] = services_binary.sum(axis=1)

#### Average Charge per Service

df['AvgChargePerService'] = df['MonthlyCharges'] / (df['ServiceAdoptionScore']+1)

df

df.to_csv('data/processed/4_feature_engineered.csv', index=False)
