==> 5_feature_binning.ipynb <==
### Import Dependencies

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.base import BaseEstimator, TransformerMixin

### Loading the Data

df = pd.read_csv('data/processed/4_feature_engineered.csv')
df

### Feature Binning

bins = [0, 12, 48, float('inf')]
labels = ['New', 'Established', 'Loyal']

df['TenureBins'] = pd.cut(df['tenure'], bins=bins, labels=labels, right= True, include_lowest=True)


df['PaymentReliabilityIndicator'] = df['PaymentMethod'].map({'Electronic check': 'Unreliable', 'Mailed check': 'Unreliable' , 'Bank transfer (automatic)': 'Reliable', 'Credit card (automatic)': 'Reliable'})

df.drop(columns=['tenure', 'PaymentMethod'], inplace=True)

df

df.to_csv('data/processed/5_feature_binned.csv', index=False)
