==> 6_feature_encoding.ipynb <==
### Import Dependencies

import os
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt

### Loading the Data

df = pd.read_csv('data/processed/5_feature_binned.csv')

nominal_cols = ["gender", "SeniorCitizen", "Partner", "Dependents",
                "PhoneService", "MultipleLines", "InternetService",
                "OnlineSecurity", "OnlineBackup", "DeviceProtection",
                "TechSupport", "StreamingTV", "StreamingMovies",
                "PaperlessBilling", "PaymentReliabilityIndicator"]

ordinal_cols = ["Contract", "TenureBins"]

df_encoded = df.copy()
df_encoded

### Nominal Columns

for col in nominal_cols:
    dummies = pd.get_dummies(df[col], prefix=col,drop_first=True)
    df_encoded = pd.concat([df_encoded, dummies], axis=1)
    df_encoded.drop(col, axis=1, inplace=True)

df_encoded

### Ordinal Columns


df_encoded['TenureBins'] = df_encoded['TenureBins'].map({'New': 0, 'Established': 1, 'Loyal': 2})
df_encoded['Contract'] = df_encoded['Contract'].map({'Month-to-month': 0, 'One year': 1, 'Two year': 2})

df_encoded.to_csv('data/processed/6_feature_encoded.csv', index=False)
