==> 7_feature_scaling.ipynb <==
### Import Dependencies

import os
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
from sklearn.preprocessing import StandardScaler, MinMaxScaler

pd.options.display.max_columns = None

### Loading the Data

df = pd.read_csv('data/processed/6_feature_encoded.csv')
df

columns_to_scale = ['MonthlyCharges', 'TotalCharges', 'AvgChargePerService']

for col in columns_to_scale:
    scaler = StandardScaler()
    df[col] = scaler.fit_transform(df[[col]])

df

df.to_csv('data/processed/7_feature_scaled.csv', index=False)
