==> 10_multi_model_training.ipynb <==
### Import Dependencies

import warnings
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
from sklearn.linear_model import LogisticRegression
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier
from xgboost import XGBClassifier
from catboost import CatBoostClassifier
from sklearn.metrics import (
                            classification_report, 
                            confusion_matrix,
                            roc_auc_score,
                            roc_curve,
                            precision_recall_curve,
                            f1_score,
                            precision_score,
                            recall_score,
                            accuracy_score
                            )
from sklearn.model_selection import (
                            StratifiedKFold,
                            cross_validate,
                            GridSearchCV,
                            RandomizedSearchCV
                            )

### Loading the Data

X_train = np.load('artifacts/X_train.npz', allow_pickle=True)['arr_0']
y_train = np.load('artifacts/y_train.npz', allow_pickle=True)['arr_0']
X_test = np.load('artifacts/X_test.npz', allow_pickle=True)['arr_0']
y_test = np.load('artifacts/y_test.npz', allow_pickle=True)['arr_0']

### Models

models = {
    'Random Forest': RandomForestClassifier(),
    'XGBoost': XGBClassifier(),
    'Cat Boost': CatBoostClassifier()
}

### Parameter Grids

rf_param_grid = {
    'n_estimators': [100, 200, 300],
    'max_depth': [6, 8, 12, 16],
    'criterion': ['gini', 'entropy', 'log_loss'],
    'min_samples_split': [2, 5, 10],
    'min_samples_leaf': [1, 2, 4]
}


xgb_param_grid = {
    'n_estimators': [100, 200, 250, 300],
    'max_depth': [3, 4, 6, 8, 10],
    'learning_rate': [0.01, 0.05, 0.1, 0.2],
    'subsample': [0.6, 0.8, 1.0],
    'colsample_bytree': [0.6, 0.8, 1.0],
    'gamma': [0, 0.1, 0.2]
}


cat_param_grid ={
    'iterations': [200, 500, 1000],
    'depth': [2, 4, 6, 8, 10],
    'learning_rate': [0.01, 0.05, 0.1, 0.2],
    'l2_leaf_reg': [1, 3, 5, 7],
    'border_count': [32, 64, 128]
}


param_grids = {
    'Random Forest' : rf_param_grid,
    'XGBoost': xgb_param_grid,
    'Cat Boost': cat_param_grid 
}

### Configure K-Fold CV

cv = StratifiedKFold(
    n_splits=5,
    random_state=42,
    shuffle=True
)

from sklearn.model_selection import StratifiedKFold, RandomizedSearchCV

cv = StratifiedKFold(
    n_splits=5,
    random_state=42,
    shuffle=True
)

random_search_results = {}

for model_name, model in models.items():
    print(f'\nTuning {model_name}')
    
    # param_distributions instead of param_grid
    param_dist = param_grids[model_name]  
    
    random_search = RandomizedSearchCV(
        estimator=model,
        param_distributions=param_dist,
        n_iter=20,            
        cv=cv,
        scoring='f1',
        verbose=1,
        random_state=42,
        n_jobs=-1                
    )

    print(f'Fitting RandomizedSearchCV for {model_name}...')
    random_search.fit(X_train, y_train)
    random_search_results[model_name] = random_search

    print(f'\n{model_name} RandomizedSearchCV completed')
    print(f'Best parameters : {random_search.best_params_}')
    print(f'Best CV score : {random_search.best_score_}')


model_cat = CatBoostClassifier(
    learning_rate = 0.05,
    l2_leaf_reg = 3,
    iterations = 500,
    depth = 10,
    border_count = 64
)

model_cat.fit(X_train, y_train)

y_hat_cat = model_cat.predict(X_test)

accuracy = accuracy_score(y_test, y_hat_cat)
precision = precision_score(y_test, y_hat_cat)
recall = recall_score(y_test, y_hat_cat)
f1 = f1_score(y_test, y_hat_cat)

print(f'accuracy: {accuracy}')
print(f'precision: {precision}')
print(f'recall: {recall}') 
print(f'f1_score: {f1}')
